+++ src/DTO/Config/HandlerConfig/FileHandlerConfig.php
@@ -1,0 +1,8 @@
+<?php
+
+namespace DTO\Config\HandlerConfig;
+
+readonly class FileHandlerConfig implements HandlerConfigInterface
+{
+
+}