+++ src/DTO/Config/HandlerConfig/HandlerConfigInterface.php
@@ -1,0 +1,10 @@
+<?php
+
+namespace DTO\Config\HandlerConfig;
+
+interface HandlerConfigInterface
+{
+    const string TYPE_FILE = 'file';
+    const string TYPE_FORWARD = 'forward';
+    const string TYPE_REDIRECT = 'redirect';
+}