+++ src/Exception/ConfigException.php
@@ -1,0 +1,9 @@
+<?php
+
+namespace Exception;
+
+use RuntimeException;
+
+class ConfigException extends RuntimeException
+{
+}