+++ CHANGELOG.md
@@ -3,6 +3,16 @@
 All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning v2.0.0](http://semver.org/).
 
+## [0.2.0] - 2026-01-23
+### Added
+- Added hosts configuration using XML files
+- Added request routing based on regular expressions
+- Added redirect and request forwarding handlers
+- Added logging
+
+### Fixed
+- Fixed a bug with data stream transmission when generating a server response
+
 ## [0.1.0] - 2026-01-20
 ### Added
 - First web server release